Enforce timezone to be set during test
During the forbidden names sniff test for correct usage of keywords
the file must have no errors. If the date.timezone ini setting
has not been set, it will fail the test. Add some code to set the timezone
if it has not been set to avoid the failure message