Encapsulated a load of duplicated code into a base class, which the individual sniffs extend.
Previously, each class was repeating the same rather lengthy version check, to decide whether a sniff error should be reported (based on the specified testVersion), which makes the code hard to read, increases the risk of errors when editing (particularly when new sniffs are added) and makes the logic harder to change in the future (e.g. to specify a range of versions that you need your code to support).

I have therefore created a base class, called PHPCompatibility_Sniff, which all the other sniffs in the project now extend.  This implements PHP_CodeSniffer_Sniff, so child classes no longer need to mention this.  It adds two helper methods, easily available to all sniffs:

* supportsAbove($phpVersion) - returns true if the version being tested against is >= $phpVersion or if no testVersion was specified.
* supportsBelow($phpVersion) - returns true if the version being tested against is <= $phpVersion (returns false if no testVersion was specified).

The logic is identical to the two tests that were previously duplicated all over the place, but it is now encapsulated in a much more easy-to-manage way.

All unit-tests pass: OK (245 tests, 955 assertions)