Fixed false positives from being reported by NewKeywordsSniff.php.
This occurred with code that includes a keyword being used as a function name, when checking against a version of PHP where this is not a problem, e.g. before the keyword was introduced.  We already have the FrobiddenNamesAsInvokedFunctionsSniff which checks for incorrect uses of keywords as functions (based on the testVersion), but we were also getting a false-positive for some valid uses of keywords as functions.

For example, the code "goto();" with testVersion = 5.2 should not generate any errors, but it was generating the following: "goto" keyword is not present in PHP version 5.2 or earlier

This fixes the unit test that was broken by 8444db0, which failed as follows:

  1) ForbiddenNamesAsInvokedFunctionsSniffTest::testGotoInPreviousVersion
  Failed asserting no standards violation on line 12: ERROR: "goto" keyword is not present in PHP version 5.2 or earlier

The test was failing because I added T_GOTO to the list of keywords that this NewKeywordsSniff checked, but the ForbiddenNamesAsInvokedFunctionsSniffTest had a test to ensure that goto() was valid in PHP < 5.3.  The false-positive, which I have now fixed, caused this test to fail.