Following-on from 06e6010, I've fixed another test that was failing due to false-positives introduced by the new checks added in 8444db0.
These failed because they assumed no error would be raised if a reserved keyword is used as a class name in a version of PHP before it became reserved.  However, the new check raised an error as the sniff didn't take into account situations where it was being used in this manner.

I have now added the check, so the false-positive is avoided.