Removing T_STATIC from NewKeywordsSniff.php
In 64bca88 I added some missing keywords to NewKeywordsSniff.php.  For some reason, this included a check for T_STATIC in PHP < 5.3.  I have no idea where I got this from - this keyword has been around since PHP 4 and its use for class methods has been allowed since PHP 5.0.

The only possibility I can think of is that there was an error in the PHP documentation (http://php.net/manual/en/reserved.keywords.php) as that is what I was checking against when I logged ticket #45.  If that was the case, then the documentation has since been fixed, as there is no trace of it now.  Either that or I was just going mad!