I've added a helper function for unit-testing, called requiresPHPVersion() to the BaseSniffTest class, and updated code to use it.
This function marks the test as skipped if it is running on a version of PHP for which the test will incorrectly fail (e.g. because the tokenizer won't generate the correct tokens).

Previously there were some tests which used a version_compare() to bypass their functionality on unsupported versions, and others where the body of the test was simply commented out (so it wasn't running on any PHP version) and in both cases the result is that these tests incorrectly appear to pass.  I have updated these tests to use the new function instead, so they are now correctly reported as skipped, which gives a more accurate reflection of the passing state of the test-suite.