Tidying of the way we retrieve the testVersion property.
Rather than repeatedly calling PHP_CodeSniffer::getConfigData('testVersion'), we now access the value via a getTestVersion() method, which caches the result of this call, meaning the config is only accessed once.

In itself, I have no idea whether this will improve performance (it depends on how getConfigData() works) but it makes the code clearer/simpler, and it allows for some enhancements that I have planned, which will involve parsing/validating the config string and which we don't want to have to duplicate, or to perform multiple times.