Fix for issue #42 (PHP notice error relating to 'nested_parenthesis' array index).
> Logs periodically show messages like the following while scanning a large tree:
> PHP Notice: Undefined index: nested_parenthesis in
> ~Sniffs/PHP/ForbiddenCallTimePassByReferenceSniff.php on line 114

The issue is relating to the 'nested_parenthesis' property not being set in the token information in some situations.

This may be related to the following CodeSniffer bug: https://pear.php.net/bugs/bug.php?id=9274.  This indicates that there are a few token types do not include this property, but it is not clear whether this is something that was fixed (in which case this is only an issue for old versions) or wontfixed, in which case current versions are likely to still be affected.  It is also possible that that issue is not relvant to this one at all.

Either way, this commit fixes the problem by checking that the array key exists before doing anything with it.