Review of visibility of all sniff properties and removing of some superfluous ones.
As with the readme change in #130, the ability to change public properties from a custom ruleset will become more well-known, it seemed prudent to verify that all `public` properties are properties which should be user-adjustable.

As most were not, I've changed the visibility of most `public` properties to `protected`.

Also a `public $error` property was available in a lot of classes, but was either not used at all or only used by one method in these classes and overruled within that method no matter what a user might have passed via a custom ruleset, so I've removed these properties in favour of a local method variable.