Allow for classes which *extend* potentially newly introduced classes.
The `NewClasses` sniff check for instantiation of classes `new DateTime;`, but did not check for classes extending other classes, i.e. `class MyOwnDateTime extends DateTime {}`.

This PR fixes that. Includes unit tests.

As `JsonSerializable` is an interface and not a class, this is not tested with the `extends` syntax. In reality, `JsonSerializable` should probably be moved to a `NewInterfaces` sniff.