Replace a lot of near duplicate unit test functions with data providers.
Makes maintenance of the unit tests easier if something would change and makes adding new unit tests a breeze.

This change also makes the unit tests more complete as often only the 'error' case was tested and not the 'noViolation' case. This has now been rectified.

The number of tests being run will be now also be reported more accurately.

For this PR I have focused on the test files where the most advantage could be gained from using data providers.

Additional changes:
* Added a constant for the test file name and used that throughout the tests in the files I touched.
* Add proper skip test indication for tests being skipped based on PHPCS version.
* Renamed the `NewAnonymousClasses` test file (contained a typo).
* Show deprecation for removed `ereg` extension and show `pcre` as an alternative. (`DeprecatedExtensions`)
* Show removal for `ereg` and `mysql` functions. (`DeprecatedFunctions`).
* Show deprecation for `mysqli` functions. (`DeprecatedFunctions`).
* Corrected some version numbers. (`DeprecatedFunctions`).
* Removed a few `setUp()` functions which did nothing but fall through to the parent.
* A number of `noViolation` unit test cases for the `NonStaticMagicMethods` sniff existed, but most were not actually being tested. That's now corrected.
* Some documentation improvements.

This is a first iteration for refactoring of the unit tests.
A further abstraction of a lot of these unit tests is definitely possible in the future.