Fix issue with running unit tests against current PHPCS master.
For some reason which I haven't been able to figure out, PHPCS master still throws an error when `quiet` is not defined and running with a test suite.

This small fix, gets rid of the error and will allow the test suite to run against master again.

The only PHPCS version which *will* give an issue is 2.6.2 which is missing part of the commit for the addition of the quiet parameter in PHPCS.