Add unit tests for the NewFunctions sniff.
The functions in the `NewFunctions` sniff appeared not to be covered by unit tests.
These have now been added.

Also:
* Make function name comparison case-insensitive as function calls are case-insensitive in PHP.
* Exclude namespaced functions from this sniff (basic check).
* Removed empty item from functions list
* Added numerous extra functions to sniff for.

Additionally some extra `noViolation` tests have been set up to make sure namespaced function calls will not be reported by this sniff.

Removed the new methods which were mixed into the functions list as they can not be tested for in this way. Most cannot be called statically and if not called statically, there is little way to tell if the method belongs to that class.