Feature completeness `NewScalarTypeDeclarations` sniff (code review).
This adds a number of additional checks to the `NewScalarTypeDeclarations` sniff.
* Check that type hints are not being used in pre-PHP 5.0 code.
* Check for valid usage of the `array` and `callable` type  hints.
* Check for invalid type hints which are typical mistakes.
* Check that the `self` type hint is only used within the class scope.

Also:
* The `description` key in the `$newTypes` array did not actually add any value, so I've removed it.
* Minor tweaks to the original error message text.
   Was along the lines of: _"int type is not present in PHP version 5.6 or earlier"_.
   Now: _"'int' type declaration is not present in PHP version 5.6 or earlier"_.

Includes unit tests covering all changes.

This PR also adds two new utility methods to the abstract `PHPCompatibility_Sniff` class.
One is a copy of an existing method from PHPCS with some minor tweaks which have been send in as a PR to PHPCS. That PR is expected to be merged for the 2.7.0 release. Unit tests for this method are available within the PHPCS test suite.
The other is a new PHPCompatibility native method `inClassScope` and is accompanied by unit tests for the method.

Similarly to an earlier PR, the file/sniff name does now no longer cover what the sniff actually does. File and sniff renaming advised.