Add newMagicMethods sniff.
Closes #64.

This sniff checks for the definition of magic methods and warns that their magic behaviour will not be invoked prior to the version it was introduced in.

For the `__toString()`  method it has a distinct message which describes the changed behaviour of the magic method between PHP 5.1 and 5.2

Includes unit tests.
Similar to PR #174, the unit test cases are split into two files - one covering classes and interfaces, one covering traits. The trait related unit tests will be skipped on PHPCS 1.x combined with PHP 5.3 or lower.