Add section about testing other sniff libs to readme.
I was running into issues when locally testing other sniff libs which do use the PHPCS native testing framework.
The short of it was that this would result in a fatal error before any unit tests would be run as PHPCS was trying to load the PHPCompatibility unit test and sniff files and the setup is differently from what PHPCS expects.

I've figured out a solution for this which resulted in PR https://github.com/squizlabs/PHP_CodeSniffer/pull/1146

Once that PR is merged, the loading of the PHPCompatibility unit tests can be bypassed allowing for the testing of other sniff libraries without running into problems.

This PR adds a section to the Readme explaining how to apply the bypass for other developers who work on both PHPCompatibility as well as other sniff libs.