Fix error message version check in the `NewIniDirectives` sniff.
The error message was showing in one version too high.
I.e. `mail.log` ini directives was introduced in PHP 5.3. The error message *should* show for PHP 5.2 and lower, not for PHP 5.3, but in practice it was also showing for PHP 5.3.

Also:
* Adjusted the actual error message to be in line with this. Adjusted the message to mirror similar messages used in other sniffs.
* For ini directives introduced in patch versions of PHP, adjusted the _"last version before"_ in the data array to be in line with the adjusted version check.
* Changed the error message to use the native PHPCS variable replacement.
* Changed the error message **type** to `error` when the ini directive was used in an `ini_set()`, left it as `warning` for `ini_get()`. This is also how it is done in the `DeprecatedIniDirectives` sniff, so these two are now in line with each other.

The unit tests have also been adjusted to account for this change.