Fix weird undefined index bug in `NewKeywords` sniff.
This bug seems to only manifest during unit testing, but would cause unit tests to fail on a `Undefined index "T_OPEN_PARENTHESIS"` in the `addError()` method of this sniff.

As the sniffing is quite token specific, it is weird that a `T_OPEN_PARENTHESIS` token would ever reach that method. All the same, the notice is interfering with other tests, so we might as well prevent it being thrown.

This fixes that.