Add new cross-PHPCS-version `inUseScope()` utility method.
This method handles checking whether a token is in a T_USE scope and works around limitations of the PHPCS 1.x implementation where `use {}` did not yet create a scope condition.

This method should be favoured over directly calling `tokenHasScope()` for checking whether a token is in a T_USE scope.

Includes unit tests for the new utility method.