Allow for an array of regex patterns in the PregReplaceEModifierSniff
The first parameter of a `preg_replace()` function call can be either a string or an array.
Up to now, the sniff always treated the first parameter as a string, leading to false positives (and potentially missing messages for regexes within an array).

Now, each individual entry in the array will be analysed and passed along for inspection.