Backport the NewAnonymousClassesSniff for older PHPCS versions.
This seemed like a quick fix to make.

In PHPCS 2.3.4 - 2.5.2, this sniff would throw an error about the `T_ANON_CLASS` constant not being defined (as it wasn't fully implemented properly in PHPCS until PHPCS 2.6.0).
Pre-PHPCS 2.3.4, the sniff was disregarded and the test skipped.

Includes additional test against false positives (test case was already in place, just not tested).
Removes test exclusion for this sniff on PHPCS <2.3.4.