Allow for array parameters in a function call used in a define.
If a function call is made to set the value of a constant through `define()`, it depends on the function called whether or not the value passed to `define()` is one of the acceptable variable types, but determining what the variable type returned by the function call is (or could be), is outside of the scope of a static analyse tool.

So I believe, in that case we should not throw an error.

Includes additional unit tests.
Also changes the `testVersion` for the `noViolation` / false positives test to `5.6` to prevet interference from the short array syntax sniff.