Tests: stabilize utility function unit tests
Most utility function tests were currently based on an **exact** token position in the test file.
This made adding/changing unit tests unnecessarily complicated and made the tests very error-prone.

I've now added a helper function to the `MethodTestFrame` class to help get the token position being targeted for a unit test based on comments in the test case file.
The principle of this is similar to how these type of functions are being tested in PHPCS upstream.

This will make these unit tests more stable, more reliable and a **lot** easier to add to/adjust.

Notes:
* Includes one extra unit test case for `testGetFQClassNameFromNewToken()`.
* Includes two extra unit test cases for `testGetFQExtendedClassName()`.

* The test for the `getFunctionCallParameterCount()` function has been split off into its own file (+ test case file).
* The test case file for the `testGetFunctionCallParameters()` and the `testGetFunctionCallParameter()` tests has been reduced to only the test actually used by those functions and the related test dataproviders have been adjusted to use offset from the stackPtr instead of exact position to allow for the tests to work with finding the target token based on a comment.