Tests: Add noViolation tests for complete test case files where appropriate.
Remove line based noViolation/version tests from test case function in favour of a file based noViolation/version test.

We don't need to test individual lines for no violation if we can test that no violations are found at all in the complete file.

This will lower the number of assertions which will be run for each build without compromising the test accuracy.

This was previously not possible because of bleed through of the results of other compatibility sniffs. As, once this PR is merged, the test case files will be tested only against the sniff for which they are intended, this becomes possible.

For version based sniffs, a complete file noViolation test is added with a version below the first deprecation/removal /above the first addition.