Tests: Always test against a testVersion. Fix missing or incorrect versions.
Related to issue 142 - not providing a testVersion could lead to incorrect test results.

Only exception is tests covering code which doesn't use the `testVersion` config variable and/or the `supportsAbove()` / `supportsBelow()` methods. In that case, the exception has been documented.

Also reviewed the version nrs used for the following:
* Version nr used in the `noFalsePositives` test should be the same as the version nr used in the "do we get the intended errors?" test.
* For `noFalsePositives` tests for complex version array based sniffs:
     - for new features, use a low version - below the first newly added feature -
     - for deprecated/removed feature, use `99.0` as the `testVersion` to be future proof.

Includes removing of some `$_sniffFile` properties which were being set without a version nr.

N.B.: Retrieving the file with the same `testVersion` several times is not an issue as the files are being cached in the test parent class.