NewNowdocQuotedHeredoc: Add sniffing for quoted heredoc identifiers.
As the logic needed is nearly the same - especially for the recognition of these tokens in PHP 5.2 -, I've added sniffing for quoted heredoc identifiers to the nowdoc sniff and renamed the sniff to match.
As the nowdoc sniff was only recently introduced and only contained in the most recent release, this should not cause a significant backward-compatibility break.

Includes unit tests.
Includes improved documentation.