Add anonymous class support in the InternalInterfaces and NewInterfaces sniffs.
While there have no new interfaces been introduced since the introduction of anonymous classes, this will make the sniff forward compatible for when those are introduced.
For InternalInterfaces, the change is already relevant.

Includes unit tests.

As there are now more tests, largely reworked the InternalInterfaces unit tests to data providers.
In a similar vain, reworked the `testUnsupportedMethods` test in the `NewInterfaces` unit test file to use a data provider.