ForbiddenNames: allow for the `final` keyword in trait use statements
According to the RFC which introduced traits, the `final` modifier is supported as well. The sniff did so far not account for this and would throw a false positive for use of the reserved keyword `final` in that case.

Example code:
```php
use sayHello as final;
use sayHello as final myFinalHello;
```

Ref: https://wiki.php.net/rfc/horizontalreuse#visibility

Includes a new usecase test file.