Issue #333: add cross-version support for anonymous classes to the RemovedGlobalVariable sniff.
Prevent false positives for Class properties named after removed PHP superglobals in PHPCS 2.4.0-2.7.1.

Pre-PHPCS 2.4.0, the `T_ANON_CLASS` token was recognized as T_CLASS and supported as such which works perfectly for our purposes.

PHPCS 2.4.0 - 2.7.1 supported the T_ANON_CLASS *token*, but did not consistently take it into account in all the relevant methods, like `getMemberProperties()`.

This was finally fixed in PHPCS 2.8.0.

As there is another upcoming sniff which will run into the same issue, I've abstracted the code to determine whether a variable is a class property to a method in the `PHPCompatibility_Sniff`.

Includes unit tests both for anonymous class support in the RemovedGlobalVariable sniff, as well as for the new utility method.