Fix the build.
As PHPCS 3.0 has been released and become the `master` branch, the builds will fail.

This PR changes the "high" branch to build against to `2.9.x-dev`/`2.9` which is the patch branch for the 2.x series and limits the "last stable" build to the 2.x series.

It also adds a note to the Readme annotating that PHPCS 3.x is currently not (yet) supported.