NewKeywords sniff: use the new `isClassConstant()` utility method.
This gives more precise results on when to throw an error and when not to for the `const` keyword.

Changes the `condition` index in the array from an array of token scopes to a call back function. As `const` is so far the only one with a condition, this has no impact on the other keywords.

Includes additional unit tests.