:fire: Hot fix: using unbounded ranges in `testVersion` only showed first error.
The overwriting of `$testVersion` within the function created a second entry with `array(null,null)` as value. This in effect broke the function result caching.

This resulted in the very first sniff calling the `supportsBelow()` method to report an error, but any subsequent calls to `supportBelow()` - even from the same sniff - would fail causing errors not to be reported.