Updating version numbers in NewClassSniff for some early Exception classes.
According to the PHP documentation (http://php.net/manual/en/class.exception.php) the Exception class was added in PHP 5.1 and our tests reflect that.  In addition, there were a number of other child classes that the documentation says were introduced in 5.0 but which our tests flagged as 5.1 as they descend from Exception and therefore couldn't be logically introduced before that class.

I have therefore updated these tests to reflect the fact that all these classes were introduced in PHP 5.0 (despite what the manual says).  I have tested all of them to confirm this, except for mysqli_sql_exception which is provided by an extension that wasn't available on my test machine.  However, as it follows the same M.O. as the others I think it's safe to assume it is correct to change this, too.

Tests were also updated.

Fixes #440.