:books: (temporarily) make it very clear that PHPCS 3.x is not (yet) supported
While I'm nearly ready with the PHPCS 3.x compatible version, it cannot be pulled/released yet until a new stable version of PHPCS 3 has been released as the latest stable version contains a number of insurmountable bugs.
So, for now, make it explicit that PHPCS 3.x is not (yet) supported.

* Also: in `CONTRIBUTING` PHPUnit 3.x was referenced, while PHPCS itself recommends PHPUnit 4 and this standard can deal with PHPUnit 4, 5, 6.