PHP 7.0/NewKeywordsSniff: sniff for `yield from`.
`yield from` turns out to be tokenized as a distinct new token `T_YIELD_FROM` which was previously now accounted for in PHPCompatibility.

Some additional logic needed to be added to deal with distinguishing between the `yield` and `yield from` keywords as we now have to deal with overlapping and multi-token keywords.

Fixes 476