PHPCS 3.x compat: Make the test bootstrap process compatible with PHPCS 3.x
In PHPCS 3.x, PHPCS does not allow autoloading via the Composer autoload file, so we also need to include the PHPCS native one.

While the autoloading of the PHPCompatibility abstract classes works fine during a normal PHPCS run, it runs into trouble when used during unit testing.
This is fixed by registering an - unit test only - autoloader for PHPCompatibility.
Upstream this is expected to be fixed in PHPCS 3.1 and the work-around could be removed once the minimum PHPCS version for the PHPCompatibility standard icm with PHPCS 3.x is raised to 3.1.

Other than that, some constants are expected in PHPCS 3.x and the file name for the file containing the PHPCS native autoloader has changed.