PHPCS 3.x compat: Account for upstream PHPCS bug
While PHPCS 1.x and 2.x allow for using namespaced sniffs, the functionality to change the settings inline using a `@codingStandardsChangeSetting` was not compatible with this until PHPCS 2.6.0.

The upgrade to support PHPCS 3.x means that PHPCompatibility now uses namespaces, so this effectively means that the custom `functionWhitelist` property will no longer function in combination with PHPCS < 2.6.0.

The related unit tests will now only be run for PHPCS >= 2.6.0 and this functional change has been annotated in the Readme section about the custom property.