:rocket: Fix the travis builds for changes in the travis infrastructure
Travis CI announced they are switching to Trusty as the default distro starting July 18th and of course, this is impacting our builds.
See: https://blog.travis-ci.com/2017-07-11-trusty-as-default-linux-is-coming

The net effect is that PHP 5.3 will only be available on `precise`, but they have also announced that they will remove that option starting September 2017, so I'm not sure what we can do from then on.

For now, this should buy us some time.

The change over to `trusty` apparently also means that `libxml2` is no longer available by default.
This was breaking the build for the XML code style check jobs. By setting it as one of the addon packages we require, this should be fixed again.

Last point I'm addressing here is PHP 5.4 in combination with PHPCS 1.x and low PHPCS 2.x versions. For completely unfathomable reasons, this build was not passing anymore with lots of unit test failures, even though it shouldn't really be affected by the above mentioned changes.
Even so, moving that build to `precise` as well, fixes it.