Hotfix: temporary plaster to fix upstream issue
When the `installed_paths` variable is set via the ruleset and not using `--config-set` on the command line, PHPCS has trouble loading arbitrary files for external rulesets, like the main `Sniff` class.

Discovered in the WordPress Coding Standards, but the same issue applies to PHPCompatibility as well.
See: WordPress-Coding-Standards/WordPress-Coding-Standards/issues/ 1087

The underlying issue has been reported upstream squizlabs/PHP_CodeSniffer/ 1591 and should be fixed there.

In the mean time, this will fix it for the time being.

I believe that this fix warrants a release of 8.0.1 soon after.