Sync `FindExtendedClassname()` method with upstream
Synced in the changes as were just merged from squizlabs/PHP_CodeSniffer/pull/ 1473

This allows the method to also detect the "Extended Classname" for interfaces extending other interfaces.

Adjusted the related `getFQExtendedClassName()` method accordingly and added unit tests for this change.

The change to the `FindExtendedClassname()` method is tested upstream.