Travis: run the build tests against PHP 7.2 as well
The new Trusty images as per Sept 7 include an image for PHP 7.2 (even though it hasn't been released yet) and nightly is now PHP 7.3-dev.

Also, while the Travis images in Q2 required PHP 5.4 to be run on `precise` for the builds to pass, seems that in Q3, `trusty` is fixed again.