ForbiddenNamesAsInvokedFunctions: don't throw error on method invocations / PHP 7.0+
How the PHP and PHPCS tokenizers tokenize the function name in a method call varies wildly across versions, sometimes its tokenized as its own token, often times its tokenized as `T_STRING`.
In practice this meant that method invocations pre-PHP 7.0 were not always caught by this sniff, while they should be.
At the same time, for PHP 7.0+, those should be exempt for the purposes of this sniff as the keyword reservation has been relaxed to allow these keywords to be used as method names.

Both these issues are fixed by this PR.

Fixes 515