NewLanguageConstructs: Update the version number for T_COALESCE_EQUAL
While this language construct is already tokenized correctly in PHPCS, PHP itself has not yet implemented it.

The RFC has been approved a while back and it was slated to go into PHP 7.2, but as of PHP 7.2 RC4 no implementation has been merged and it is not expected to go into PHP 7.2 anymore.