PHP 7.2: New OptionalRequiredFunctionParameters sniff
Previously a `RequiredOptionalFunctionParameters` sniff has been added which checks for missing function call parameters which were required and only became optional in a later PHP version.

This PR adds a sister-sniff which checks for missing function call parameters which originally were optional, but later became required.

Initially this sniff check for the `$results` parameter for the `parse_str()` function for which the optional nature of the parameter became deprecated in PHP 7.2 (soft required).
It is expected that the parameter will become hard required in PHP 8.0.

Includes unit tests.

Notes:
The commit contains a few minor changes to the `RequiredOptionalFunctionParameters` sniff to bring both sniffs in line with each other.