MethodTestFrame: make the `getTargetToken()` method more flexible and precise
The `getTargetToken()` method is used to find a token in a test case file based on a preceding comment.

In case the comment was not found, this method could inadvertently return a stackPtr to a completely different token, which could - by accident, not design - pass the test.

In case the target token was not found before the next `/* Case ... */` comment, the method could inadvertently return a stackPtr to the token for a next case, which could - by accident, not design - pass the test.

Both these cases have now been addressed.

Additionally, the method now allows for optionally passing a `$tokenContent` parameter to target one specific token in a code snippet which contains several tokens of the same type.