:sparkles: New `NewConstants` sniff to detect usage of new PHP constants
The list of constants is based on the various migration guides, the changelog and the reserved constants list in the PHP Manual.

As the `T_STRING` token is used as a kind of "catch-all", there may initially be false positives for code structures I did not consider. All the same, chances are small as that would also require the `T_STRING` to have the same content and case as the PHP native constant name.

This sniff does not automatically detect back-fills, but the error codes are modular, so if a PHP constant back-fill is in place in a certain project, that message for that particular constant can be excluded from the custom ruleset.

Includes unit tests.

The sniff also does not detect clashes between aliased constants and the PHP constants, i.e. `use const ABC as NEW_CONSTANT;`.
It would be an interesting potential future enhancement to detect these kind of aliases and see if we can on-the-fly exclude the current file being examined for that particular error code.