Build/PHPCS: check array indentation
PHPCS upstream has added a new sniff to verify normalized array indentation.
As the PHPCS build check for the PHPCompatibility standard is run against `master`, we can start taken advantage of that.

This commit adds the relevant changes to the PHPCompatibility own ruleset to do so and fixes the one instance where the code did not yet comply with this.