ForbiddenCallTimePassByReference: fix false positive for "bitwise and" preceded by global constant
This was a known false positive and annotated as such in the unit test files.

Simplified the `isCallTimePassByReferenceParam()` method by using the upstream `isReference()` function to determine whether a _bitwise and_ is a reference or not.
The upstream method has some issues pre PHPCS 3.1.1, but none which impact this sniff.

 Fixes 39