Sniff: add new `Sniff::getReturnTypeHintName()` method
For single token return type declarations, the existing `Sniff::getReturnTypeHintToken()` method is sufficient to figure out whether a return type is used and what it is.

However for class and interface based return declarations which may be preceeded by a namespace with namespace separators, just having access to the classname is not sufficient.

This new method allows to retrieve the full name of the return type declaration in a manner which is PHPCS cross-version compatible.