Test bootstrap: make the builds faster
The new Xdebug version 2.6.0 will have a feature to speed up code coverage a lot.
This code snippet enables that feature when code coverage is being run and the right Xdebug version is detected.

At this moment, the only Travis image which comes with Xdebug 2.6.0(beta1) is PHP 7.2.
Initial tests show a speed up of 60% in run-time for the unit tests on those images with this snippet in place.

Refs:
* https://twitter.com/derickr/status/953268665255940103
* https://github.com/sebastianbergmann/php-code-coverage/issues/514#issuecomment-354654930
* https://gist.github.com/hollodotme/1becd28a52a75ea4a120790ca262bf44