New `ReservedFunctionNames`sniff
This was previously handled by an upstream sniff - see PR 173 -.

The "problem" with the previous implementation is that when PHPCompatibility is used in a custom ruleset which also includes sniffs from other rulesets, a "conflict" would be created if another ruleset included the `Generic.NamingConventions.CamelCapsFunctionName` sniff as well.

PHPCompatibility silenced two of the error messages from that sniff and unless you are actually aware of that and know how to re-enable those, chances are that this will go unnoticed and non-camelCaps function and method name issues would go unreported because of it.
This breaking of other rulesets, is, of course, undesired.

This is now solved by the introduction of a new sniff which extends the upstream sniff to allow it to benefit from the upstream properties and logic.
The upstream sniff is no longer included via the ruleset, the superfluous error messages are no longer excluded, which effectively prevents the above described issue.

Included unit tests.