NewScalarTypeDeclarations: fix incorrect results
Based on a variety of tests run on 3v4l, I've come to the conclusion that this sniff needs adjusting.

* `parent` is a valid parameter type declaration.
* `self` and `parent` were not recognized properly until PHP 5.2. Before that, they would be treated as class names.

This PR adjusts the sniff to deal correctly with this based on these finding.

Includes adjusted unit tests.

N.B.: While there has previously been discussion on whether or not the "debug"-type error for `self used outside of class scope` was warranted, the above mentioned tests showed that as of PHP 7.0, using `self`/`parent` as a type declaration outside class scope results in a fatal error.
I've documented this inline.