PregReplaceEModifier: allow the sniff to work with preg_replace_callback_array()
The `preg_replace_callback_array()` function expects an array as the first parameter with the patterns as the keys and the callbacks as values.

For all other regex functions, if the first parameter is an array, the patterns are in the values.

This commit adjusts the PregReplaceEModifier sniff to allow it to also work with `preg_replace_callback_array()`.