:books: README: remove PHP "not tested" badge
As PHP 5.2 is not supported anymore anyway, it is confusing and misleading to display it as being "not tested".

Other than that, the current "tested" badge does not reflect that the library is tested against PHP 7.2 (and nightly and HHVM).

I've [asked the maintainer of PHP-Eye a question about this on Twitter](https://twitter.com/jrf_nl/status/972472140913049600). Hopefully that will be fixed soon.

[skip ci]