Sniff: add three utility methods to handle number recognition
These methods can be used to determine if variable assignments, function call parameters or array values would be regarded as a (positive/negative) number by PHP.

The `isNumber()` method is used to determine whether something would be regarded as a number by PHP.
This method takes type juggling from string and boolean to integer/float into account.

The method will return `false` if something isn't a number as well as when it couldn't be reliably determined, so should only be used for testing whether something definitely *is* a number (well, that is _would be regarded as a number by PHP_).

The method will return `false` when it encounters calculations, even when all parts of the calculation are numbers. Handling this even better, is something which could possibly be added later (in a future PR). For now, number recognition will already be greatly improved when using these utility methods.

The additional utility methods `isPositiveNumber()` and `isNegativeNumber()` are mostly there to prevent duplicate code in individual sniffs.

Includes extensive unit tests.

N.B.: There are two new sniffs upcoming in which these methods will be used, aside from the bugfix for the negative bitshift sniff.